Update miniconda_env_script.sh
Added QoL features and updated content added to .bashrc on install.

+++ miniconda_env_script.sh
@@ -3,14 +3,18 @@
 
 envname=PLUTOenv
 
-mkdir -p ~/miniconda3
-wget https://repo.anaconda.com/miniconda/Miniconda3-latest-Linux-x86_64.sh -O ~/miniconda3/miniconda.sh
-bash ~/miniconda3/miniconda.sh -b -u -p ~/miniconda3
-rm -rf ~/miniconda3/miniconda.sh
-
-/home/${USER}/miniconda3/bin/conda init
-
-source /home/${USER}/.bashrc
+if [ ! -d "/home/${USER}/miniconda3" ] ; then
+        mkdir -p ~/miniconda3
+        wget https://repo.anaconda.com/miniconda/Miniconda3-latest-Linux-x86_64.sh -O ~/miniconda3/miniconda.sh
+        bash ~/miniconda3/miniconda.sh -b -u -p ~/miniconda3
+        rm -rf ~/miniconda3/miniconda.sh
+        
+        /home/${USER}/miniconda3/bin/conda init
+        
+        source /home/${USER}/.bashrc
+else
+        echo "Miniconda appears to already be installed"
+fi
 
 #make sure .bashrc is sourced automatically on login
 if [ ! -f /home/${USER}/.profile ] ; then
@@ -22,6 +26,11 @@ conda create --name ${envname} --file spec-file.txt
 #fix missing .so.1 files
 ln -s /home/${USER}/miniconda3/envs/${envname}/lib/libmpi_cxx.so.40.20.1  /home/${USER}/miniconda3/envs/${envname}/lib/libmpi_cxx.so.1
 ln -s /home/${USER}/miniconda3/envs/${envname}/lib/libmpi.so.40.20.2  /home/${USER}/miniconda3/envs/${envname}/lib/libmpi.so.1
+
+echo -e "\n\n" >> /home/${USER}/.bashrc
+echo 'export PLUTO_DIR=/home/${USER}/PLUTO' >>  /home/${USER}/.bashrc
+echo 'export PYTHONPATH=/home/tmk5/PLUTO/Tools/pyPLUTO/:${PYTHONPATH}' >> /home/${USER}/.bashrc
+echo 'alias plutosetup="python3 ${PLUTO_DIR}/setup.py"' >> /home/${USER}/.bashrc
 
 source /home/${USER}/.bashrc
